Build: move dependency transpilation config to calypso-build
Alternative to #33492 which still uses the special Babel config for transpiling dependencies,
but moves it to the `calypso-build` package.

This PR, unlike #33492, doesn't attempt to transpile both app and `node_modules` code with the
same Babel config. It keeps the separation and shouldn't cause any changes in output -- just
moves code to a different location.

Also, this approach is in line with what Create React App does.

After merging this, `calypso-build` should start transpiling dependencies out of the box, without
any configuration. Jetpack can just upgrade to the new version and be done.
