+++ packages/calypso-build/babel/dependencies.js
@@ -1,4 +1,4 @@
-const config = {
+module.exports = () => ( {
 	// see https://github.com/webpack/webpack/issues/4039#issuecomment-419284940
 	sourceType: 'unambiguous',
 	presets: [
@@ -25,6 +25,4 @@ const config = {
 			},
 		],
 	],
-};
-
-module.exports = config;
+} );
