+++ packages/calypso-build/webpack.config.js
@@ -20,7 +20,7 @@ const DependencyExtractionWebpackPlugin = require( '@wordpress/dependency-extrac
 /**
  * Internal dependencies
  */
-const { cssNameFromFilename } = require( './webpack/util' );
+const { cssNameFromFilename, shouldTranspileDependency } = require( './webpack/util' );
 // const { workerCount } = require( './webpack.common' ); // todo: shard...
 
 /**
@@ -110,6 +110,12 @@ function getWebpackConfig(
 					configFile: babelConfig,
 					exclude: /node_modules\//,
 					presets,
+					workerCount,
+				} ),
+				TranspileConfig.loader( {
+					cacheDirectory: true,
+					include: shouldTranspileDependency,
+					presets: [ path.join( __dirname, 'babel', 'dependencies' ) ],
 					workerCount,
 				} ),
 				SassConfig.loader( { postCssConfig: { path: postCssConfigPath } } ),
